Tim1 moved to encoders.h
