+++ MainBoardODR/Core/Inc/encoders.h
@@ -16,6 +16,10 @@
 uint16_t timerValue;
 uint32_t encoderA1, encoderA2, encoderA3, encoderB1, encoderB2, encoderB3;
 
+TIM_HandleTypeDef htim1;
+
+void MX_TIM1_Init(void);
+
 void encoderInit();
 uint32_t encoderCalculateSpeed(uint32_t time, uint32_t timePrev);
 uint32_t encoderCalculateOverallSpeed(uint32_t speed1, uint32_t speed2, uint32_t speed3);
